requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.2
django-autofixture==0.12.1
django-crispy-forms==1.6.1
django-tables2==1.7.1
django-templateaddons3==1.0
django-watson==1.3.1
django-widget-tweaks==1.4.1
djangorestframework==3.6.3
gunicorn==19.7.1
olefile==0.44
Pillow==4.1.1
psycopg2==2.7.1
python-dateutil==2.6.0
python-decouple==3.0
pytz==2017.2
six==1.10.0
static3==0.7.0
whitenoise==3.3.0
